requirements.txt
pydantic~=2.11.7
python-dotenv~=1.1.1
openai~=1.99.6